+++ common/db/src/main/java/is/codion/common/db/operation/DatabaseProcedure.java
@@ -28,8 +28,8 @@ import is.codion.common.db.exception.DatabaseException;
 public interface DatabaseProcedure<C, T> {
 
   /**
-   * Executes this procedure with the given connection
-   * @param connection the connection being used to execute this procedure
+   * Executes this procedure using the given connection
+   * @param connection the connection to use
    * @param argument the procedure argument, if any
    * @throws DatabaseException in case of an exception during the execution
    */