+++ common/db/src/main/java/is/codion/common/db/operation/DefaultProcedureType.java
@@ -18,8 +18,6 @@
  */
 package is.codion.common.db.operation;
 
-import is.codion.common.db.exception.DatabaseException;
-
 import java.io.Serializable;
 
 import static java.util.Objects.requireNonNull;
@@ -32,11 +30,6 @@ final class DefaultProcedureType<C, T> implements ProcedureType<C, T>, Serializa
 
   DefaultProcedureType(String name) {
     this.name = requireNonNull(name, "name");
-  }
-
-  @Override
-  public void execute(C connection, DatabaseProcedure<C, T> procedure, T argument) throws DatabaseException {
-    requireNonNull(procedure, "procedure").execute(connection, argument);
   }
 
   @Override