+++ common/db/src/main/java/is/codion/common/db/operation/FunctionType.java
@@ -18,8 +18,6 @@
  */
 package is.codion.common.db.operation;
 
-import is.codion.common.db.exception.DatabaseException;
-
 /**
  * @param <C> the connection type
  * @param <T> the function argument type
@@ -31,16 +29,6 @@ public interface FunctionType<C, T, R> {
    * @return the function name
    */
   String name();
-
-  /**
-   * Executes the given function.
-   * @param connection the connection being used
-   * @param function the function to execute
-   * @param argument the function argument, if any
-   * @return the function result
-   * @throws DatabaseException in case of an exception
-   */
-  R execute(C connection, DatabaseFunction<C, T, R> function, T argument) throws DatabaseException;
 
   /**
    * Creates a {@link FunctionType} with the given name and types.