+++ common/db/src/test/java/is/codion/common/db/operation/DefaultFunctionTypeTest.java
@@ -18,8 +18,6 @@
  */
 package is.codion.common.db.operation;
 
-import is.codion.common.db.exception.DatabaseException;
-
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -27,10 +25,9 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public final class DefaultFunctionTypeTest {
 
   @Test
-  void test() throws DatabaseException {
+  void test() {
     DefaultFunctionType<String, String, String> functionType = new DefaultFunctionType<>("test");
     assertEquals(functionType, new DefaultFunctionType<>("test"));
     assertEquals(functionType.name(), functionType.toString());
-    functionType.execute("conn", (connection, argument) -> "", "hello");
   }
 }