+++ common/model/src/main/java/is/codion/common/model/table/TableSelectionModel.java
@@ -175,7 +175,7 @@ public interface TableSelectionModel<R> {
   List<Integer> getSelectedIndexes();
 
   /**
-   * Selects all visible entities
+   * Selects all visible rows
    * @see #addSelectionListener(Runnable)
    */
   void selectAll();