+++ framework/db-rmi/src/main/java/is/codion/framework/db/rmi/RemoteEntityConnection.java
@@ -247,7 +247,7 @@ public interface RemoteEntityConnection extends Remote, AutoCloseable {
 
   /**
    * Performs an update based on the given update, updating the columns found
-   * in the {@link Update#columnValues()} map, with the associated values.
+   * in the {@link Update#values()} map, using the associated value.
    * @param update the update to perform
    * @return the number of affected rows
    * @throws DatabaseException in case of a dabase exception