+++ framework/domain/src/main/java/is/codion/framework/domain/DefaultDomain.java
@@ -267,6 +267,7 @@ public abstract class DefaultDomain implements Domain {
     private final Map<ProcedureType<?, ?>, DatabaseProcedure<?, ?>> procedures = new HashMap<>();
 
     private void addProcedure(ProcedureType<?, ?> procedureType, DatabaseProcedure<?, ?> procedure) {
+      requireNonNull(procedureType, "procedureType");
       requireNonNull(procedure, "procedure");
       if (procedures.containsKey(procedureType)) {
         throw new IllegalArgumentException("Procedure already defined: " + procedureType);
@@ -291,6 +292,7 @@ public abstract class DefaultDomain implements Domain {
     private final Map<FunctionType<?, ?, ?>, DatabaseFunction<?, ?, ?>> functions = new HashMap<>();
 
     private void addFunction(FunctionType<?, ?, ?> functionType, DatabaseFunction<?, ?, ?> function) {
+      requireNonNull(functionType, "functionType");
       requireNonNull(function, "function");
       if (functions.containsKey(functionType)) {
         throw new IllegalArgumentException("Function already defined: " + functionType);