+++ framework/json-db/src/test/java/is/codion/framework/json/db/DatabaseObjectMapperTest.java
@@ -101,7 +101,7 @@ public final class DatabaseObjectMapperTest {
     Update readCondition = mapper.readValue(jsonString, Update.class);
 
     assertEquals(update.where(), readCondition.where());
-    assertEquals(update.columnValues(), readCondition.columnValues());
+    assertEquals(update.values(), readCondition.values());
   }
 
   @Test